refactor(proxy): consolidate thinking plugins into ClaudeRequestNormalizerPlugin
Merged haiku-thinking-disabler and thinking-transformer into single plugin.
Extracted transformation logic to standalone handler functions outside plugin class.

Changes:
- Renamed to ClaudeRequestNormalizerPlugin (was ClaudeThinkingTransformerPlugin)
- Chain pattern: handleNoThinkingModels → handleAdaptiveThinkingTransform
- Updated regex to cover haiku 3-5 and 4-5 (excludes 4-6+)
- Removed duplicate haiku-thinking-disabler plugin
- Priority 14 (before RequestSanitizer at 15)
