+++ src/providers/plugins/sso/proxy/plugins/index.ts
@@ -12,10 +12,9 @@ import { SSOAuthPlugin } from "./sso-auth.plugin.js";
 import { JWTAuthPlugin } from "./jwt-auth.plugin.js";
 import { HeaderInjectionPlugin } from "./header-injection.plugin.js";
 import { RequestSanitizerPlugin } from "./request-sanitizer.plugin.js";
-import { ClaudeThinkingTransformerPlugin } from "./claude-thinking-transformer.plugin.js";
+import { ClaudeRequestNormalizerPlugin } from "./claude-request-normalizer.plugin.js";
 import { LoggingPlugin } from "./logging.plugin.js";
 import { SSOSessionSyncPlugin } from "./sso.session-sync.plugin.js";
-import { ClaudeThinkingDisablerPlugin } from "./haiku-thinking-disabler.plugin.js";
 
 /**
  * Register core plugins
@@ -26,12 +25,11 @@ export function registerCorePlugins(): void {
 
   // Register in any order (priority determines execution order)
   registry.register(new MCPAuthPlugin()); // Priority 3 - MCP auth relay routing
-  registry.register(new ClaudeThinkingDisablerPlugin()); // Priority 3 - MCP auth relay routing
   registry.register(new EndpointBlockerPlugin()); // Priority 5 - blocks unwanted endpoints early
   registry.register(new SSOAuthPlugin());
   registry.register(new JWTAuthPlugin());
+  registry.register(new ClaudeRequestNormalizerPlugin()); // Priority 14 - normalizes thinking params for claude models
   registry.register(new RequestSanitizerPlugin()); // Priority 15 - strips unsupported reasoning params
-  registry.register(new ClaudeThinkingTransformerPlugin()); // Priority 16 - transforms thinking params for Claude 4.7+ models
   registry.register(new HeaderInjectionPlugin());
   registry.register(new LoggingPlugin()); // Always enabled - logs to log files at INFO level
   registry.register(new SSOSessionSyncPlugin()); // Priority 100 - syncs sessions via multiple processors
@@ -48,8 +46,7 @@ export {
   JWTAuthPlugin,
   HeaderInjectionPlugin,
   RequestSanitizerPlugin,
-  ClaudeThinkingTransformerPlugin,
-  ClaudeThinkingDisablerPlugin,
+  ClaudeRequestNormalizerPlugin,
   LoggingPlugin,
 };
 export { SSOSessionSyncPlugin } from "./sso.session-sync.plugin.js";
